Use external worker rather than blob worker so that we can control its source policy.
This also requires splitting scriptWorker and scriptEditor so that we can reuse the worker.
